requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.1

# --- Environment ---
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# --- AI / LLM & Transcription ---
numpy==1.26.4
torch==2.1.2
openai-whisper==20231117
openai==1.42.0
groq

# --- Audio Utils ---
ffmpeg-python==0.2.0

# --- Utilities ---
requests==2.32.3
boto3==1.35.2

# --- Dev Tools ---
pytest==8.3.3
httpx==0.27.2
black==24.8.0
isort==5.13.2
